+++ cli/srecli/service/constants.py
@@ -64,7 +64,7 @@ class Endpoint(str, Enum):
     BATCH_RESULTS = '/batch-results'
     REPORTS_RETRY = '/reports/retry'
     POLICIES_NAME = '/policies/{name}'
-    SERVICE_OPERATION_STATUS = '/service-operation/status'
+    SERVICE_OPERATIONS_STATUS = '/service-operations/status'
     REPORTS_CLEVEL = '/reports/clevel'
     METRICS_UPDATE = '/metrics/update'
     METADATA_UPDATE = '/metadata/update'
@@ -260,7 +260,10 @@ class JobState(str, Enum):
         return map(operator.attrgetter('value'), cls)
 
 
-OPERATION_STATUS_HINT = "Use 'sre service_operation status --operation {operation_type}' to check execution status"
+OPERATION_STATUS_HINT = (
+    "Use 'sre service operations status --operation {operation_type}' "
+    "to check execution status."
+)
 
 
 class ServiceOperationType(Enum):
@@ -271,7 +274,7 @@ class ServiceOperationType(Enum):
     sre service_operation status --operation {service_operation_type} (snake_case)
 
     For the API endpoint, use the following format:
-    /service-operation/status?type={service-operation-type} (kebab-case)
+    /service-operations/status?type={service_operation_type} (kebab-case)
     """
 
     UPDATE_METRICS = ('metrics_update', 'metrics-update')